Merged PR 97274: Fixed baseline text through trial and error
Fixed baseline text through trial and error

This is done with a nice spreadsheet
And verfied with which magical constant works best
![image.png](https://dev.azure.com/trackman/582e98b1-b4a1-48b7-a943-3c5189477c5e/_apis/git/repositories/96e3c40a-7904-405b-bab4-9b0a0665d846/pullRequests/97274/attachments/image.png)
After testing these configurations of two texts in UI Toolkit. It showed that 1/-2.75 offset was the most correct.

The previous value assumed 1/-5 was correct when text was bottom aligned. But didnt factor in lineheight.
By setting text to center aligned on the counteraxis, it ignores the lineheight. Which makes this emulation much more stable.   But then we needed a new constant

I guess it makes sense thatthe new value is apprx half the other. Since we are now setting text to be centered isntaed of bottom.

Example. Baseline text in figma of size 20 and 15
Adds -2 to bottom of the text element of 15 size
![image (4).png](https://dev.azure.com/trackman/582e98b1-b4a1-48b7-a943-3c5189477c5e/_apis/git/repositories/96e3c40a-7904-405b-bab4-9b0a0665d846/pullRequests/97274/attachments/image%20%284%29.png)
![image (5).png](https://dev.azure.com/trackman/582e98b1-b4a1-48b7-a943-3c5189477c5e/_apis/git/repositories/96e3c40a-7904-405b-bab4-9b0a0665d846/pullRequests/97274/attachments/image%20%285%29.png)
Using two different fonts cannot be emulated as in figma yet. And offset might be a few pixels

Related work items: #248293