Merged PR 99681: Shared styles from figma shouldnt take alignment and wrapping.
Shared styles from figma shouldnt take alignment and wrapping. Since that is not what the shared style defines.

In current version we added all style properties for the shared style. This is font size, weight, type, wrapping, alignment and truncation.
But the style only defines font style, weight and type

![image.png](https://dev.azure.com/trackman/582e98b1-b4a1-48b7-a943-3c5189477c5e/_apis/git/repositories/96e3c40a-7904-405b-bab4-9b0a0665d846/pullRequests/99681/attachments/image.png)

Solved this by making two methods. One that only gets the font-style, weight and type.

It results in this change
![image (2).png](https://dev.azure.com/trackman/582e98b1-b4a1-48b7-a943-3c5189477c5e/_apis/git/repositories/96e3c40a-7904-405b-bab4-9b0a0665d846/pullRequests/99681/attachments/image%20%282%29.png)

Had difficulty understanding the flow of shared styles, so I renamed some variables. Key And Value to StyleType and StyleId

Related work items: #249862