Merged PR 114889: Suppressing IOException when we are trying to write the same element multiple...
Resolves #261005

Whenever we inherit a class in a multiple classes with Query in a base class that has a Template we end up in a situation, when the element of the Template is being requested n-times you inherit the class.

So in order to avoid having (possibly) the same template written multiple times and having IOException, we do not include a template that already has the same name.

Related work items: #261005